services/agent-service/requirements.txt
grpcio>=1.60.0
grpcio-health-checking>=1.60.0
langchain>=0.1.0
langchain-google-genai>=0.0.10
python-decouple>=3.8
